+++ src/BUILD_
@@ -10,6 +10,7 @@ go_binary(
         '//src/cli',
         '//src/core',
         '//src/export',
+        '//src/follow',
         '//src/gc',
         '//src/hashes',
         '//src/help',
@@ -17,7 +18,6 @@ go_binary(
         '//src/output',
         '//src/parse',
         '//src/query',
-        '//src/remote',
         '//src/run',
         '//src/test',
         '//src/tool',
