+++ src/follow/proto/BUILD
@@ -2,5 +2,5 @@ grpc_library(
     name = 'build_event',
     srcs = ['build_event.proto'],
     languages = ['go'],
-    visibility = ['//src/remote/...'],
+    visibility = ['//src/follow/...'],
 )
